Legger til api for å endre status på eksisterende oppgave.
Flytter DTO klasser til kontrakt modul.
