--- app/src/main/kotlin/no/nav/ung/deltakelseopplyser/domene/deltaker/DeltakerDTO.kt
@@ -1,21 +1,0 @@
-package no.nav.ung.deltakelseopplyser.domene.deltaker
-
-import java.util.*
-
-data class DeltakerDTO(
-    val id: UUID ? = null,
-    val deltakerIdent: String,
-) {
-    companion object {
-        fun DeltakerDTO.mapToDAO(): DeltakerDAO {
-            return DeltakerDAO(deltakerIdent = deltakerIdent)
-        }
-
-        fun DeltakerDAO.mapToDTO(): DeltakerDTO {
-            return DeltakerDTO(
-                id = id,
-                deltakerIdent = deltakerIdent
-            )
-        }
-    }
-}
