+++ app/src/main/kotlin/no/nav/ung/deltakelseopplyser/domene/oppgave/repository/OppgavetypeDataDAO.kt
@@ -2,10 +2,31 @@ package no.nav.ung.deltakelseopplyser.domene.oppgave.repository
 
 import com.fasterxml.jackson.annotation.JsonFormat
 import com.fasterxml.jackson.annotation.JsonProperty
-import no.nav.ung.deltakelseopplyser.domene.oppgave.OppgavetypeDataJsonType
+import com.fasterxml.jackson.annotation.JsonSubTypes
+import com.fasterxml.jackson.annotation.JsonTypeInfo
+import no.nav.ung.deltakelseopplyser.kontrakt.oppgave.felles.ArbeidOgFrilansRegisterInntektDTO
+import no.nav.ung.deltakelseopplyser.kontrakt.oppgave.felles.RegisterinntektDTO
+import no.nav.ung.deltakelseopplyser.kontrakt.oppgave.felles.YtelseRegisterInntektDTO
 import java.time.LocalDate
 
-@OppgavetypeDataJsonType
+@JsonTypeInfo(
+    use = JsonTypeInfo.Id.NAME,
+    include = JsonTypeInfo.As.PROPERTY,
+    property = "type"
+)
+@JsonSubTypes(
+    // Endret startdato oppgavetype data
+    JsonSubTypes.Type(value = EndretStartdatoOppgavetypeDataDAO::class, name = "BEKREFT_ENDRET_STARTDATO"),
+
+    // Endret sluttdato oppgavetype data
+    JsonSubTypes.Type(value = EndretSluttdatoOppgavetypeDataDAO::class, name = "BEKREFT_ENDRET_SLUTTDATO"),
+
+    // Kontroller registerinntekt oppgavetype data
+    JsonSubTypes.Type(
+        value = KontrollerRegisterInntektOppgaveTypeDataDAO::class,
+        name = "BEKREFT_AVVIK_REGISTERINNTEKT"
+    )
+)
 sealed class OppgavetypeDataDAO
 
 data class EndretStartdatoOppgavetypeDataDAO(
@@ -15,7 +36,7 @@ data class EndretStartdatoOppgavetypeDataDAO(
 ) : OppgavetypeDataDAO()
 
 data class EndretSluttdatoOppgavetypeDataDAO(
-    @JsonFormat(pattern = "yyyy-MM-dd")val nySluttdato: LocalDate,
+    @JsonFormat(pattern = "yyyy-MM-dd") val nySluttdato: LocalDate,
     @JsonProperty(defaultValue = "n/a") val veilederRef: String,
     val meldingFraVeileder: String?,
 ) : OppgavetypeDataDAO()
@@ -29,21 +50,33 @@ data class KontrollerRegisterInntektOppgaveTypeDataDAO(
 
     @JsonFormat(pattern = "yyyy-MM-dd")
     @JsonProperty(defaultValue = "n/a") val tomDato: LocalDate,
-    ) : OppgavetypeDataDAO()
+) : OppgavetypeDataDAO()
 
 
 data class RegisterinntektDAO(
     @JsonProperty("arbeidOgFrilansInntekter") val arbeidOgFrilansInntekter: List<ArbeidOgFrilansRegisterInntektDAO>,
-    @JsonProperty("ytelseInntekter") val ytelseInntekter: List<YtelseRegisterInntektDAO>
-)
+    @JsonProperty("ytelseInntekter") val ytelseInntekter: List<YtelseRegisterInntektDAO>,
+) {
+    companion object {
+        fun RegisterinntektDAO.tilDTO() = RegisterinntektDTO(
+            arbeidOgFrilansInntekter = arbeidOgFrilansInntekter.map {
+                ArbeidOgFrilansRegisterInntektDTO(
+                    it.inntekt,
+                    it.arbeidsgiver
+                )
+            },
+            ytelseInntekter = ytelseInntekter.map { YtelseRegisterInntektDTO(it.inntekt, it.ytelsetype) }
+        )
+    }
+}
 
 data class ArbeidOgFrilansRegisterInntektDAO(
     @JsonProperty("inntekt") val inntekt: Int,
-    @JsonProperty("arbeidsgiver") val arbeidsgiver: String
+    @JsonProperty("arbeidsgiver") val arbeidsgiver: String,
 )
 
 
 data class YtelseRegisterInntektDAO(
     @JsonProperty("inntekt") val inntekt: Int,
-    @JsonProperty("ytelsetype") val ytelsetype: String
+    @JsonProperty("ytelsetype") val ytelsetype: String,
 )
