+++ app/src/main/kotlin/no/nav/ung/deltakelseopplyser/domene/register/ungsak/UngdomsprogramRegisterK9SakController.kt
@@ -5,9 +5,8 @@ import io.swagger.v3.oas.annotations.tags.Tag
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import no.nav.security.token.support.core.api.RequiredIssuers
 import no.nav.ung.deltakelseopplyser.config.Issuers
-import no.nav.ung.deltakelseopplyser.domene.register.DeltakerOpplysningerDTO
+import no.nav.ung.deltakelseopplyser.kontrakt.register.DeltakerOpplysningerDTO
 import no.nav.ung.deltakelseopplyser.domene.register.UngdomsprogramregisterService
-import no.nav.ung.deltakelseopplyser.integration.pdl.api.PdlService
 import no.nav.ung.sak.kontrakt.person.AktørIdDto
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
