+++ kontrakt/src/main/kotlin/no/nav/ung/deltakelseopplyser/kontrakt/deltaker/DeltakerDTO.kt
@@ -1,0 +1,8 @@
+package no.nav.ung.deltakelseopplyser.kontrakt.deltaker
+
+import java.util.*
+
+data class DeltakerDTO(
+    val id: UUID ? = null,
+    val deltakerIdent: String,
+)
