+++ kontrakt/src/main/kotlin/no/nav/ung/deltakelseopplyser/kontrakt/deltaker/RapportPeriodeinfoDTO.kt
@@ -1,18 +1,7 @@
-package no.nav.ung.deltakelseopplyser.domene.register
+package no.nav.ung.deltakelseopplyser.kontrakt.deltaker
 
-import no.nav.ung.deltakelseopplyser.domene.oppgave.OppgaveDTO
 import java.math.BigDecimal
 import java.time.LocalDate
-import java.util.*
-
-data class DeltakelsePeriodInfo(
-    val id: UUID,
-    val fraOgMed: LocalDate,
-    val tilOgMed: LocalDate? = null,
-    val harSøkt: Boolean,
-    val oppgaver: List<OppgaveDTO>,
-    val rapporteringsPerioder: List<RapportPeriodeinfoDTO> = emptyList(),
-)
 
 data class RapportPeriodeinfoDTO(
     val fraOgMed: LocalDate,
