+++ kontrakt/src/main/kotlin/no/nav/ung/deltakelseopplyser/kontrakt/oppgave/felles/OppdaterOppgaveStatusDTO.kt
@@ -1,0 +1,5 @@
+package no.nav.ung.deltakelseopplyser.kontrakt.oppgave.felles
+
+data class OppdaterOppgaveStatusDTO(
+    val status: OppgaveStatus
+)
