+++ kontrakt/src/main/kotlin/no/nav/ung/deltakelseopplyser/kontrakt/oppgave/felles/OppgaveStatus.kt
@@ -1,0 +1,7 @@
+package no.nav.ung.deltakelseopplyser.kontrakt.oppgave.felles
+
+enum class OppgaveStatus {
+    LØST,
+    ULØST,
+    AVBRUTT,
+}
