+++ kontrakt/src/main/kotlin/no/nav/ung/deltakelseopplyser/kontrakt/oppgave/felles/Oppgavetype.kt
@@ -1,0 +1,7 @@
+package no.nav.ung.deltakelseopplyser.kontrakt.oppgave.felles
+
+enum class Oppgavetype {
+    BEKREFT_ENDRET_STARTDATO,
+    BEKREFT_ENDRET_SLUTTDATO,
+    BEKREFT_AVVIK_REGISTERINNTEKT,
+}
