+++ kontrakt/src/main/kotlin/no/nav/ung/deltakelseopplyser/kontrakt/oppgave/felles/OppgavetypeDataJsonType.kt
@@ -1,11 +1,8 @@
-package no.nav.ung.deltakelseopplyser.domene.oppgave
+package no.nav.ung.deltakelseopplyser.kontrakt.oppgave.felles
 
 import com.fasterxml.jackson.annotation.JacksonAnnotationsInside
 import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
-import no.nav.ung.deltakelseopplyser.domene.oppgave.repository.EndretSluttdatoOppgavetypeDataDAO
-import no.nav.ung.deltakelseopplyser.domene.oppgave.repository.EndretStartdatoOppgavetypeDataDAO
-import no.nav.ung.deltakelseopplyser.domene.oppgave.repository.KontrollerRegisterInntektOppgaveTypeDataDAO
 
 /**
  * Denne annotasjonen brukes for polymorf serialisering og deserialisering
@@ -27,14 +24,14 @@ import no.nav.ung.deltakelseopplyser.domene.oppgave.repository.KontrollerRegiste
 @JsonSubTypes(
     // Endret startdato oppgavetype data
     JsonSubTypes.Type(value = EndretStartdatoOppgavetypeDataDTO::class, name = "BEKREFT_ENDRET_STARTDATO"),
-    JsonSubTypes.Type(value = EndretStartdatoOppgavetypeDataDAO::class, name = "BEKREFT_ENDRET_STARTDATO"),
 
     // Endret sluttdato oppgavetype data
     JsonSubTypes.Type(value = EndretSluttdatoOppgavetypeDataDTO::class, name = "BEKREFT_ENDRET_SLUTTDATO"),
-    JsonSubTypes.Type(value = EndretSluttdatoOppgavetypeDataDAO::class, name = "BEKREFT_ENDRET_SLUTTDATO"),
 
     // Kontroller registerinntekt oppgavetype data
-    JsonSubTypes.Type(value = KontrollerRegisterinntektOppgavetypeDataDTO::class, name = "BEKREFT_AVVIK_REGISTERINNTEKT"),
-    JsonSubTypes.Type(value = KontrollerRegisterInntektOppgaveTypeDataDAO::class, name = "BEKREFT_AVVIK_REGISTERINNTEKT")
+    JsonSubTypes.Type(
+        value = KontrollerRegisterinntektOppgavetypeDataDTO::class,
+        name = "BEKREFT_AVVIK_REGISTERINNTEKT"
+    ),
 )
 annotation class OppgavetypeDataJsonType
