+++ kontrakt/src/main/kotlin/no/nav/ung/deltakelseopplyser/kontrakt/veileder/DeltakelseInnmeldingDTO.kt
@@ -1,4 +1,4 @@
-package no.nav.ung.deltakelseopplyser.domene.register
+package no.nav.ung.deltakelseopplyser.kontrakt.veileder
 
 import java.time.LocalDate
